Added some code + refactoring

+++ src/HomeLibrary/service/runner/RunAsAdmin.java
@@ -6,9 +6,9 @@ import HomeLibrary.service.LibraryLogic;
 import HomeLibrary.entity.user.User;
 import HomeLibrary.entity.user.Users;
 import HomeLibrary.view.Print;
-import HomeLibrary.view.View;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Scanner;
 
 public class RunAsAdmin {
@@ -23,15 +23,16 @@ public class RunAsAdmin {
         do {
             Print.printTheMessage("" +
                     "1 - Watch library, " +
-                    "2 - Add book to the library, " +
-                    "3 - Delete book, " +
-                    "4 - Quit");
+                    "2 - Find book in the library, " +
+                    "3 - Add book to the library, " +
+                    "4 - Delete book, " +
+                    "5 - Quit");
 
             if (scanner.hasNextInt()) {
 
                 choiceInt = scanner.nextInt();
 
-                if (choiceInt > 0 & choiceInt < 5) {
+                if (choiceInt > 0 & choiceInt < 6) {
                     check = choiceInt;
                     bool = true;
                 }
@@ -41,34 +42,50 @@ public class RunAsAdmin {
 
         switch (check) {
             case (1):
-                View.printBooks(library);
+                Print.printBooks(library);
                 runAsAdmin(users, library, user);
                 break;
 
 //--------------------------------------------------------------------------------------------------------------
             case (2):
+
+                Print.printTheMessage("What do you want to find?");
+                String textToFind = scanner.next();
+                ArrayList<Book> findingBooks = LibraryLogic.findTheBook(textToFind, library);
+                if (findingBooks == null) {
+                    Print.printTheMessage("Sorry, can't find any match like " + textToFind);
+                } else {
+                    Print.printTheMessage("Find next book(s):");
+                    Print.printTheElementsOfArrayListOfBooks(findingBooks);
+                }
+                runAsAdmin(users, library, user);
+                break;
+//--------------------------------------------------------------------------------------------------------------
+            case (3):
                 Book newBook = LibraryLogic.addNewBookToLibrary(library);
                 LibraryLogic.saveLibraryToTXT(newBook);
                 runAsAdmin(users, library, user);
                 break;
 //--------------------------------------------------------------------------------------------------------------
 
-            case (3):
-
-                View.printBooks(library);
+            case (4):
+                Print.printBooks(library);
                 Print.printTheMessage("Choose a number of book for deleting..");
 
                 int numberOfBook = scanner.nextInt();
-
-                library = LibraryLogic.deleteBookFromLibrary(numberOfBook, library);
-                LibraryLogic.saveLibraryToTXTDelete(library);
-                View.printBooks(library);
+                if (LibraryLogic.isExist(library, numberOfBook)) {
+                    library = LibraryLogic.deleteBookFromLibrary(numberOfBook, library);
+                    LibraryLogic.saveLibraryToTXTDelete(library);
+                    Print.printBooks(library);
+                } else {
+                    Print.printTheMessage("Wrong index!");
+                }
                 runAsAdmin(users, library, user);
                 break;
 
 //--------------------------------------------------------------------------------------------------------------
 
-            case (4):
+            case (5):
                 break;
 //--------------------------------------------------------------------------------------------------------------
         }
